+++ src/matchers/bool.rs
@@ -1,5 +1,4 @@
-
-use core::{Matcher, Join};
+use core::{Join, Matcher};
 
 /// A matcher for `be_true` assertions.
 pub struct BeTrue;
@@ -39,34 +38,34 @@ impl Matcher<bool, ()> for BeFalse {
 
 #[cfg(test)]
 mod tests {
-    use super::{be_true, be_false};
+    use super::{be_false, be_true};
     use core::expect;
 
     #[test]
     fn test_be_true_message() {
-        expect(1 == 0).to(be_true()).assert_eq_message(
-            "expected to be true",
-        );
+        expect(1 == 0)
+            .to(be_true())
+            .assert_eq_message("expected to be true");
     }
 
     #[test]
     fn test_not_to_be_true_message() {
-        expect(0 == 0).not_to(be_true()).assert_eq_message(
-            "expected not to be true",
-        );
+        expect(0 == 0)
+            .not_to(be_true())
+            .assert_eq_message("expected not to be true");
     }
 
     #[test]
     fn test_be_false_message() {
-        expect(0 == 0).to(be_false()).assert_eq_message(
-            "expected to be false",
-        );
+        expect(0 == 0)
+            .to(be_false())
+            .assert_eq_message("expected to be false");
     }
 
     #[test]
     fn test_not_to_be_false_message() {
-        expect(0 == 1).not_to(be_false()).assert_eq_message(
-            "expected not to be false",
-        )
+        expect(0 == 1)
+            .not_to(be_false())
+            .assert_eq_message("expected not to be false")
     }
 }