+++ src/matchers/mod.rs
@@ -1,12 +1,11 @@
-
 //! A module contains matchers.
 
 pub use self::partial_eq::{be_equal_to, BeEqualTo};
-pub use self::partial_ord::{PartialOrder, be_less_than, be_less_or_equal_to, be_greater_than,
-                            be_greater_or_equal_to};
-pub use self::bool::{be_true, BeTrue, be_false, BeFalse};
-pub use self::option::{be_some, BeSome, be_none, BeNone};
-pub use self::result::{be_ok, BeOk, be_err, BeErr};
+pub use self::partial_ord::{be_greater_or_equal_to, be_greater_than, be_less_or_equal_to,
+                            be_less_than, PartialOrder};
+pub use self::bool::{be_false, be_true, BeFalse, BeTrue};
+pub use self::option::{be_none, be_some, BeNone, BeSome};
+pub use self::result::{be_err, be_ok, BeErr, BeOk};
 pub use self::floats::{be_close_to, BeCloseTo};
 pub use self::empty::{be_empty, BeEmpty};
 pub use self::count::{have_count, HaveCount};