+++ tests/empty.rs
@@ -1,4 +1,3 @@
-
 #[macro_use(expect)]
 extern crate expectest;
 use expectest::prelude::*;