+++ CHANGELOG.md
@@ -12,6 +12,7 @@ be_less_or_equal_to <=> be_le
 be_greater_than <=> be_gt
 be_greater_or_equal_to <=> be_ge
 ```
+- Set minimum rust version in `.travis.yml`
 
 ## [0.9.2] - 2018-01-21
 - Adapted float point tests to a new Debug implementation